add ps2 non sdl

+++ src/platform/win32/win32.c
@@ -449,4 +449,4 @@ void *Platform_malloc(size_t numBytes) { return HeapAlloc(GetProcessHeap(), HEAP
 
 void Platform_free(void *ptr) { HeapFree(GetProcessHeap(), 0, ptr); }
 
-void Platform_QueueAudio(const u8 *data, u32 numBytes) { }
+void Platform_QueueAudio(const float *data, u32 numBytes) { }
